home page refactoring
